Final changes InshaaAllah

+++ app/src/main/java/com/example/liderstoreagent/data/models/clientmodel/AddClientData.kt
@@ -11,6 +11,6 @@ class AddClientData(
         val latitude: Double,
         val longitude: Double,
         val image: File ,
-        val assumptionValue: Int,
+        val assumptionValue: Long,
         val agentId: Int,
 )
